Utils folder and css add
