+++ src/components/organisms/CategoryDetail/CategoryDetail.module.css
@@ -1,5 +1,7 @@
 .main {
   width: 35%;
+  position: absolute;
+  top: 0;
 }
 
 .categories_wrapper {
@@ -66,7 +68,7 @@
 
 @media (max-width: 900px){
    .main {
-      width: 80%;
+      width: 90%;
    }
 }
