+++ src/utils/slugify.ts
@@ -1,0 +1,2 @@
+export const slugify = (text: string): string =>
+   text.toLowerCase().replace(/\s+/g, '-')
